+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ## [TODO]
 
+## [1.7.1] - 2025-08-26
+
+### Added
+- Yield result if block given to `execute` and `execute!` methods
+
 ## [1.7.0] - 2025-08-25
 
 ### Added